--- src/components/Navigation/ModeButton.scss
@@ -1,4 +1,0 @@
-.gadgetContainer {
-  display: inline-block;
-  padding: 3px 5px;
-}
