+++ src/components/Navigation/NavBar.js
@@ -1,25 +1,56 @@
 import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
+import {setModeStudent, setModeTeacher} from '../../action_creators';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import Navbar from 'react-bootstrap/lib/Navbar';
 import Nav from 'react-bootstrap/lib/Nav';
 import NavItem from 'react-bootstrap/lib/NavItem';
 //import FormControl from 'react-bootstrap/lib/FormControl';
+import MenuItem from 'react-bootstrap/lib/MenuItem';
+import DropdownButton from 'react-bootstrap/lib/DropdownButton';
 import Clearfix from 'react-bootstrap/lib/Clearfix';
 import LinkContainer from 'react-router-bootstrap/lib/LinkContainer';
 import {BreadCrumbContainer as BreadCrumb} from './BreadCrumb';
 import LanguageDropdown from './LanguageDropdown';
 import {getTranslator} from '../../selectors/translate';
 import styles from './NavBar.scss';
-import ModeButton from './ModeButton';
 
 const modes = ['student', 'teacher'];
 
-// function SearchBox(props) {
+const ModeDropdown = ({t, mode, setModeStudent, setModeTeacher}) => {
+  const texts = {'student': t('general.student'), 'teacher': t('general.teacher')};
+  const setMode = mode => mode === 'student' ? setModeStudent() : setModeTeacher();
+  return <div className={styles.gadgetContainer}>
+    <DropdownButton id='mode-dropdown'
+                    noCaret
+                    pullRight
+                    bsStyle={mode}
+                    title={t('navbar.mode') + ': ' + texts[mode]}
+                    onSelect={setMode}>
+      {
+        modes.map(k =>
+          <MenuItem key={k} eventKey={k} active={mode === k}>{texts[k]}</MenuItem>
+        )
+      }
+    </DropdownButton>
+  </div>;
+};
+
+ModeDropdown.propTypes = {
+  setModeStudent: PropTypes.func,
+  setModeTeacher: PropTypes.func,
+  mode: PropTypes.oneOf(modes).isRequired
+};
+
+// const SearchBox = ({t}) => {
 //   return <div className={styles.gadgetContainer}>
-//     <FormControl type='text' placeholder={props.t('search.placeholder')}/>
+//     <FormControl type='text' placeholder={t('search.placeholder')}/>
 //   </div>;
-// }
+// };
+
+// SearchBox.propTypes ? {
+//  t: PropTypes.func
+// };
 
 const LkkBrand = () => {
   return <Navbar.Brand>
@@ -52,16 +83,22 @@ LkkNav.propTypes = {
   t: PropTypes.func
 };
 
-const Gadgets = ({isStudentMode, t}) => {
+const Gadgets = ({isStudentMode, t, setModeStudent, setModeTeacher}) => {
   const mode = modes[isStudentMode ? 0 : 1];
+  // NOTE: Commenting out SearchBox until it is implemented
   return <div className={styles.gadgetGroup}>
     {<LanguageDropdown mode={mode}/>}
-    <ModeButton t={t} isStudentMode={isStudentMode}/>
+    <ModeDropdown setModeStudent={setModeStudent}
+                  setModeTeacher={setModeTeacher}
+                  mode={mode}
+                  t={t}/>
     {/*<SearchBox t={t}/>*/}
   </div>;
 };
 
 Gadgets.propTypes = {
+  setModeStudent: PropTypes.func,
+  setModeTeacher: PropTypes.func,
   isStudentMode: PropTypes.bool,
   t: PropTypes.func
 };
@@ -82,7 +119,7 @@ MenuToggle.propTypes = {
   t: PropTypes.func
 };
 
-export const NavBar = ({isStudentMode, t, params}) => {
+export const NavBar = ({isStudentMode, t, setModeStudent, setModeTeacher, params}) => {
   const widgetClass = isStudentMode ? styles.widgetStudent : styles.widgetTeacher;
   return (
     <div className={styles.navbarWrapper}>
@@ -98,7 +135,10 @@ export const NavBar = ({isStudentMode, t, params}) => {
         </Navbar.Collapse>
         <div className={styles.widgets + ' ' + widgetClass}>
           <BreadCrumb params={params}/>
-          <Gadgets isStudentMode={isStudentMode} t={t}/>
+          <Gadgets setModeStudent={setModeStudent}
+                   setModeTeacher={setModeTeacher}
+                   isStudentMode={isStudentMode}
+                   t={t}/>
         </div>
       </Navbar>
     </div>
@@ -111,6 +151,8 @@ NavBar.propTypes = {
     lesson: PropTypes.string,
     file: PropTypes.string
   }),
+  setModeStudent: PropTypes.func,
+  setModeTeacher: PropTypes.func,
   isStudentMode: PropTypes.bool,
   t: PropTypes.func
 };
@@ -120,6 +162,12 @@ const mapStateToProps = (state) => ({
   t: getTranslator(state)
 });
 
+const mapDispatchToProps = {
+  setModeStudent,
+  setModeTeacher
+};
+
 export const NavBarContainer = connect(
-  mapStateToProps
+  mapStateToProps,
+  mapDispatchToProps
 )(withStyles(styles)(NavBar));
